refactor(products): move product key related utils to utils module
