+++ src/controllers/products/product/productController.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from "express";
 
-import { decryptRecipeURLAndGetProductId } from "../../../services/products/helper.js";
+import { decryptRecipeURLAndGetProductId } from "../../../services/products/utils.js";
 import { validateId } from "../../../utils/numbers.js";
 import { ServiceError } from "../../../services/helpers/ServiceError.js";
 import { getProductDetail } from "../../../services/products/product/productService.js";
