+++ src/services/products/productsService.ts
@@ -1,6 +1,6 @@
 import mysqlDB from "../../db/mysql.js";
 import { lightSlugify, lightTrim } from "../../utils/normalize.js";
-import { decryptRecipeURLAndGetProductId } from "./helper.js";
+import { decryptRecipeURLAndGetProductId } from "./utils.js";
 import { formatClientProducts } from "./helper.js";
 import { Product, SearchProductsData, SearchProductsParams } from "./type.js";
 import { Ingredient } from "../ingredients/type.js";
